requirements.txt
certifi==2019.9.11
chardet==3.0.4
dicom==0.9.9.post1
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
httpie==1.0.3
idna==2.8
Markdown==3.1.1
Pillow==6.2.0
pydicom==1.3.0
Pygments==2.4.2
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
